ci(): release bump improvement (#26)
This pull request includes several updates to the release process in the
GitHub Actions configuration. The changes aim to synchronize the
`package.json` version with the latest tag, automate version
determination and updates, and ensure changes are committed and pushed
appropriately.

### Synchronization and Version Management:

*
[`.github/actions/create-release/action.yml`](diffhunk://#diff-45dad27f90da5e7332a48ea77ace56197e37c80aeadf63cf19403696145856e3L26-R55):
Added a new step to sync `package.json` version with the latest tag,
ensuring the version is up-to-date before determining the new version.
*
[`.github/actions/create-release/action.yml`](diffhunk://#diff-45dad27f90da5e7332a48ea77ace56197e37c80aeadf63cf19403696145856e3L26-R55):
Modified the version determination step to use `standard-version` for
bumping the version and updating `package.json`.

### Commit and Push Changes:

*
[`.github/actions/create-release/action.yml`](diffhunk://#diff-45dad27f90da5e7332a48ea77ace56197e37c80aeadf63cf19403696145856e3R66-R76):
Added a step to commit and push changes if a new release is needed,
ensuring the repository reflects the latest version updates.