ci(): release bump improvement (#28)
This pull request includes significant updates to the GitHub Actions
workflow for creating releases and minor changes to the CI workflow
permissions. The most important changes involve simplifying the release
creation process by using a dedicated action for version determination
and release creation, and adding outputs to the release action.

Changes to release creation workflow:

*
[`.github/actions/create-release/action.yml`](diffhunk://#diff-45dad27f90da5e7332a48ea77ace56197e37c80aeadf63cf19403696145856e3L1-R40):
Updated the action name and description, added outputs for
`release_created` and `version`, and simplified the steps by integrating
the `conventional-commits-semver-release` action for version
determination and the `release-action` for creating releases.

Changes to CI workflow permissions:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR9):
Added `contents: write` permission to the CI workflow.

Minor cleanup:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL65):
Removed an unnecessary blank line in the `jobs:` section.