ci()!: update publish-packages script (#34)
This pull request includes an update to the `publish-packages.yml`
workflow to improve the build process for packages. The key change is
the addition of error handling and reporting for package builds.

Enhancements to package build process:

* Added an `id` to the "Build All Packages" step for better
identification.
* Introduced error handling to capture and report failed package builds.
If any package fails to build, the workflow will now log the failed
packages and exit with an error.