ci: fix build order (#39)
This pull request includes changes to the GitHub Actions and workflows
to enhance the setup and build processes for the project. The most
important changes include adding ASDF paths to the environment and
splitting the build process for `@basis/bun-plugins` from the other
packages.

Enhancements to GitHub Actions setup:

*
[`.github/actions/asdf-setup/action.yml`](diffhunk://#diff-ddfd25f88e6634d4be9cc1cb6aee1b098a2d19477910edc1f144077fe26f8c9bR73-R78):
Added commands to set ASDF paths in the environment variables and update
the GitHub path.

Improvements to package build workflow:

*
[`.github/workflows/publish-packages.yml`](diffhunk://#diff-f649dfa0883378db13b96d378fbaedfaf99e75a92d0962d813575b79643dd01bL55-R77):
Split the build process to handle `@basis/bun-plugins` separately from
the other packages, ensuring better error handling and clearer build
logs.
*
[`.github/workflows/publish-packages.yml`](diffhunk://#diff-f649dfa0883378db13b96d378fbaedfaf99e75a92d0962d813575b79643dd01bL85-L87):
Removed unnecessary sourcing of `asdf.sh` script since `jq` is no longer
required in the path for the publish step.