ci!: npm-publish (#42)
This pull request includes updates to the CI and release workflows, as
well as improvements to the `getChangedFiles` function and a minor
change in the `fetchAllTags` function. The most important changes
include streamlining the CI workflow, updating the release workflow to
handle workspaces, and simplifying the logic in the `getChangedFiles`
function.

### Workflow Improvements:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL55-L60):
Streamlined the CI workflow by removing unnecessary parameters from the
`checkout` step and improving the build process for changed workspaces.
[[1]](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL55-L60)
[[2]](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR66-R83)

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L10-R11):
Updated the release workflow to identify and handle changed workspaces
instead of packages, and improved the build and publish process to skip
already existing versions.
[[1]](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L10-R11)
[[2]](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L21-L22)
[[3]](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L58-L95)

### Function Improvements:

*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L1-R27):
Simplified the `getChangedFiles` function by removing complex branching
logic and consolidating the process of fetching changes into a more
streamlined approach.

### Minor Changes:

*
[`libraries/workspaces/source/functions/fetchAllTags.ts`](diffhunk://#diff-644c6632415dd3ab75e43874f66c15fa5e190d372e6fc823c5469e8d3ee21857L5-R5):
Changed the `fetchAllTags` function to use the `.quiet()` method for a
cleaner output.