ci: no workspace references on main (#43)
This pull request includes several changes to the GitHub workflows and a
minor update to import paths in a TypeScript file. The most important
changes include refactoring the `checkout` step in workflows, adding a
log statement for changed workspaces, and updating conditions for
publishing in the release workflow.

Changes to GitHub workflows:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL56-R56):
Refactored the `checkout` step to use a more concise syntax for
`fetch-depth`.
*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR75):
Added a blank line for better readability in the script that identifies
changed workspaces.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L17-R18):
Refactored the `checkout` step to use a more concise syntax for
`fetch-depth`.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R58-R62):
Added a log statement to print the changed workspaces.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R58-R62):
Updated the condition for publishing changed workspaces to check if
there are any changes.

Changes to import paths:

*
[`libraries/workspaces/source/functions/buildPackage.ts`](diffhunk://#diff-6bf7935cd57ff57010a581f9deb1ac85356c5a4e51c19ec9cbc8521ff96cf43cL6-R7):
Updated import paths for `pluginGlobals`, `pluginSASS`, `formatBytes`,
and `formatMilliseconds` to use relative paths.