ci: npm publish (#44)
This pull request includes several changes to the CI/CD workflows and
some refactoring in the codebase. The most important changes are related
to improving the build and release processes, as well as updating import
paths for better modularity.

Improvements to CI/CD workflows:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL56-R56):
Modified the `checkout` step to use a more concise syntax for
`fetch-depth`.
*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR75):
Added a blank line after capturing the `build_output` to improve
readability.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L17-R18):
Modified the `checkout` step to use a more concise syntax for
`fetch-depth`.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R58-R62):
Added an echo statement to log the changed workspaces for better
visibility during the release process.

Code refactoring:

*
[`libraries/workspaces/source/functions/buildPackage.ts`](diffhunk://#diff-6bf7935cd57ff57010a581f9deb1ac85356c5a4e51c19ec9cbc8521ff96cf43cL6-R7):
Updated import paths for `pluginGlobals`, `pluginSASS`, `formatBytes`,
and `formatMilliseconds` to use relative paths for better modularity.
*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L16-R26):
Enhanced the logic to determine the comparison reference for `git diff`,
ensuring it correctly handles cases when at the tip of the base branch.