ci: compare to latest release when in detached head state (#49)
This pull request includes a change to the `getChangedFiles` function in
`libraries/workspaces/source/functions/getChangedFiles.ts`. The update
modifies the logic for determining the comparison reference when
checking for changed files.

Key change:

*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L16-R27):
Updated the logic to check if the current commit is at the tip of the
base branch instead of checking if the current branch is 'main'. This
change ensures that the comparison reference is determined accurately
based on the commit position rather than the branch name.