ci: get revision for tag (#54)
This pull request includes changes to improve the handling of git tags
and commit references in the `getUnreleasedCommitMessages` and
`getChangedFiles` functions. The most important changes involve updating
the logic to fetch and compare commit hashes more accurately.

Improvements to git tag handling:

*
[`.github/actions/create-release/getUnreleasedCommitMessages.ts`](diffhunk://#diff-d4b67875b7662e641c412753022c7ed1fbeededd605cf53ae79d9b355540af12L41-R47):
Updated the logic to fetch the latest tag and its commit hash for a more
accurate commit range comparison.
*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L14-R14):
Modified the logic to fetch the latest tag and its commit hash for
comparison, ensuring accurate detection of changes.

Enhancements to commit reference comparison:

*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L23-R33):
Refined the comparison logic to determine the correct reference point
based on whether the current commit is at the tip of the base branch and
whether a tag exists.