ci: properly get revision for tag (#55)
This pull request includes improvements to the way the latest tag and
its commit hash are retrieved for both unreleased commit messages and
changed files. These changes ensure more accurate comparisons and
enhance the reliability of the functions.

Improvements to tag retrieval and comparison:

*
[`.github/actions/create-release/getUnreleasedCommitMessages.ts`](diffhunk://#diff-d4b67875b7662e641c412753022c7ed1fbeededd605cf53ae79d9b355540af12L41-R47):
Updated the logic to fetch the latest tag and its commit hash, and used
this commit hash for the comparison range.
*
[`libraries/workspaces/source/functions/getChangedFiles.ts`](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L14-R14):
Modified the logic to fetch the latest tag and its commit hash for
comparison when determining changed files.
[[1]](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L14-R14)
[[2]](diffhunk://#diff-d7c98c857e069fc7a56f2b2bc42cb778b8514a5ed50464205b197d4fd5c761d8L23-R33)