ci: compute latest version bump (#62)
This pull request includes enhancements to the version determination
process and refactoring to improve code reuse in the
`.github/actions/determine-version` directory. The most important
changes include updating the `getCurrentVersion` function, adding a new
function to get the current version commit hash, and refactoring the
`getUnreleasedCommitMessages` function to use these new utilities.

Enhancements to version determination:

*
[`.github/actions/determine-version/getCurrentVersion.ts`](diffhunk://#diff-cdb49436a01124ea37f9f817ca52a264abf25a14fed7368cf752667ce66c71d7L4-R39):
Updated the `getCurrentVersion` function to unshallow the repository,
fetch all tags, and return the latest version tag. Added a new function
`getCurrentVersionCommit` to get the commit hash of the current version
tag.

Refactoring for code reuse:

*
[`.github/actions/determine-version/getUnreleasedCommitMessages.ts`](diffhunk://#diff-b144a8558ee5cb9d6599f2126e0861f3edd1f924e1cf68635d6c1728627cd55cR3):
Refactored the `getUnreleasedCommitMessages` function to use the
`getCurrentVersion` and `getCurrentVersionCommit` functions, simplifying
the process of fetching the latest tag and its commit hash.
[[1]](diffhunk://#diff-b144a8558ee5cb9d6599f2126e0861f3edd1f924e1cf68635d6c1728627cd55cR3)
[[2]](diffhunk://#diff-b144a8558ee5cb9d6599f2126e0861f3edd1f924e1cf68635d6c1728627cd55cL40-R47)