ci: JSR needs no leading v on version #'s (#64)
This pull request includes changes to improve the version determination
and display process in the GitHub Actions workflow, as well as a minor
fix in the release workflow.

Enhancements to version determination and display:

*
[`.github/actions/determine-version/action.yml`](diffhunk://#diff-64f5056f156060043dc451631e5fcb4990061c3dfe978e9ea7f9e1a766e6d255R19-R41):
Added a new step to display version information, including whether a
release is needed and the current and next version numbers.

Fixes in the release workflow:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L56-R61):
Corrected the syntax for extracting the next version from the
`determine-version` output to ensure it is properly formatted.