feat: improve Carousel (#68)
This pull request includes significant updates to the `Carousel`
component, focusing on enhancing its functionality and improving the
codebase. The changes include modifications to the component's styles,
tests, and TypeScript definitions.

### Style Improvements:
*
[`libraries/react/components/Carousel/Carousel.scss`](diffhunk://#diff-5fc53dd2881f1bcfea8da45878bd81854b6d4ef2717124d28e03e7cf292e1d7aL2-R183):
Refactored the CSS to improve the layout and visibility of navigation
elements, added hover and focus states for better user interaction, and
restructured the alignment and size modifiers for images.

### Test Enhancements:
*
[`libraries/react/components/Carousel/Carousel.test.tsx`](diffhunk://#diff-8dad8f4b85517db92d6a9ca8d437babdb133e61f605047eb6cf4ff43cae2da0fL105-L215):
Updated tests to handle new interactions such as closing the lightbox
with the escape key, responding to arrow keys when focused, and opening
images in a new tab on middle-click. Removed redundant tests and added
new ones for mouse and touch interactions.
[[1]](diffhunk://#diff-8dad8f4b85517db92d6a9ca8d437babdb133e61f605047eb6cf4ff43cae2da0fL105-L215)
[[2]](diffhunk://#diff-8dad8f4b85517db92d6a9ca8d437babdb133e61f605047eb6cf4ff43cae2da0fL224-R226)

### Codebase Simplification:
*
[`libraries/react/components/Carousel/Carousel.tsx`](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59R2):
Simplified the state management by removing the `fullSize` state and
consolidating the `data` and `attributes` getters. Added the
`createPortal` import to support rendering elements outside the main DOM
hierarchy.
[[1]](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59R2)
[[2]](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59L31-L32)
[[3]](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59L67)
[[4]](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59L76-R85)
[[5]](diffhunk://#diff-b4f47a4e1255f01f2fceedb17f469e9bd95d71d104ca763e3ea06ed4cb060c59R95-R114)